+++ kloader/crtn.s
@@ -1,16 +1,12 @@
 
 .section .init
 
-	popfl
-	popal
 	popl %ebp
 	ret
 	
 	
 .section .fini
 
-	popfl
-	popal
 	popl %ebp
 	ret
