+++ kloader/cstub.c
@@ -1,0 +1,56 @@
+
+#include "cstub.h"
+
+uint32_t memcmp(const void *a, const void *b, size_t length)
+{
+	const uint8_t *_a = (const uint8_t*)a;
+	const uint8_t *_b = (const uint8_t*)b;
+	
+	for (size_t i = 0; i < length; i++)
+	{
+		if (_a[i] < _b[i]) {
+			return -1;
+		} else if (_a[i] > _b[i]) {
+			return 1; 
+		}
+	}
+	return 0;
+}
+
+void *memcpy(void *dest, const void *src, size_t length)
+{
+	uint8_t *_target = (uint8_t*)dest;
+	for (size_t i = 0; i < length; i++)
+	{
+		_target[i] = ((uint8_t*)src)[i];
+	}
+	return dest;
+}
+
+void *memmove(void *dest, const void *source, size_t length)
+{
+	uint8_t *_target = (uint8_t*)dest;
+	if (dest < source) 
+	{
+		for (size_t i = 0; i < length; i++)
+		{
+			_target[i] = ((uint8_t*)source)[i];
+		}
+	} else {
+		for (size_t i = length; i != 0; i--)
+		{
+			_target[i-1] = ((uint8_t*)source)[i-1];
+		}		
+	}
+	return dest;
+}
+
+void *memset(void *s, int value, size_t length)
+{
+	uint8_t *_target = (uint8_t*)s;
+	for (size_t i = 0; i < length; i++)
+	{
+		_target[i] = (uint8_t) value;
+	}
+	return s;
+}
