+++ kloader/cstub.h
@@ -14,7 +14,6 @@ uint32_t memcmp(const void*, const void*, size_t);
 void* memcpy(void*, const void*, size_t);
 void* memmove(void*, const void*, size_t);
 void* memset(void*, int, size_t);
-size_t strlen(const char*);
 
 #ifdef __cplusplus
 }
