+++ kloader/main.cpp
@@ -10,11 +10,12 @@
 #include "paging.hpp"
 #include "cstub.h"
 #include "cpuid.hpp"
+#include "pmm.hpp"
 
 
 char* itoa(int num, char* str, int base);
-
-extern "C" uint32_t main(uint32_t magic, multiboot2_info_t *mb2_info)
+ 
+extern "C" uint32_t kloader_main(uint32_t magic, multiboot2_info_t *mb2_info)
 {
 	retval_t retval;
 	cpuid_retval_t cpuid_retval;
@@ -23,10 +24,12 @@ extern "C" uint32_t main(uint32_t magic, multiboot2_info_t *mb2_info)
 	char _temp_text[16];
 	memset(_temp_text, 0, 16);
 	
+	// Check Multiboot2 magic number
 	if (magic != MULTIBOOT2_INFO_MAGIC) {
 		return RETVAL_ERROR_MB2_MAGIC;
 	}
 	
+	// Init Console
 	retval = Console::Init(mb2_info);
 	if (retval != RETVAL_OK) {
 		return retval;
@@ -34,6 +37,16 @@ extern "C" uint32_t main(uint32_t magic, multiboot2_info_t *mb2_info)
 	
 	Console::Clear();
 	Console::SetTitleText("HeppOS - kloader");
+
+	// Init PMM (Physical Memory Manager)
+	Console::Print("Initializing PMM - ");
+	retval = PMM::Init(mb2_info);
+	if (retval != RETVAL_OK) {
+		Console::Print("ERROR\n");
+		return retval;
+	}
+	Console::Print("OK\n");
+	
 	
 	// Check for LongMode
 	Console::Print("Checking for LongMode - ");
@@ -190,7 +203,7 @@ extern "C" uint32_t main(uint32_t magic, multiboot2_info_t *mb2_info)
 	
 	
 	// setup initial 64bit paging (maybe map kernel space -2GB to the first 2GB in memory?)
-	Console::Print("Setup Paging - ");
+	Console::Print("Initializing Paging - ");
 	retval = Paging::Init();
 	if (retval != RETVAL_OK) {
 		Console::Print("ERROR!\n");	
@@ -200,7 +213,7 @@ extern "C" uint32_t main(uint32_t magic, multiboot2_info_t *mb2_info)
 
 	
 	// Setup GDT
-	Console::Print("Setup GDT - ");
+	Console::Print("Initializing GDT - ");
 	retval = GDT::Init();
 	if (retval != RETVAL_OK) {
 		Console::Print("ERROR\n");
