--- kloader/memcmp.c
@@ -1,18 +1,0 @@
-
-#include "cstub.h"
-
-uint32_t memcmp(const void *a, const void *b, size_t length)
-{
-	const uint8_t *_a = (const uint8_t*)a;
-	const uint8_t *_b = (const uint8_t*)b;
-	
-	for (size_t i = 0; i < length; i++)
-	{
-		if (_a[i] < _b[i]) {
-			return -1;
-		} else if (_a[i] > _b[i]) {
-			return 1; 
-		}
-	}
-	return 0;
-}
