--- kloader/memcpy.c
@@ -1,12 +1,0 @@
-
-#include "cstub.h"
-
-void *memcpy(void *dest, const void *src, size_t length)
-{
-	uint8_t *_target = (uint8_t*)dest;
-	for (size_t i = 0; i < length; i++)
-	{
-		_target[i] = ((uint8_t*)src)[i];
-	}
-	return dest;
-}
