--- kloader/memmove.c
@@ -1,20 +1,0 @@
-
-#include "cstub.h"
-
-void *memmove(void *dest, const void *source, size_t length)
-{
-	uint8_t *_target = (uint8_t*)dest;
-	if (dest < source) 
-	{
-		for (size_t i = 0; i < length; i++)
-		{
-			_target[i] = ((uint8_t*)source)[i];
-		}
-	} else {
-		for (size_t i = length; i != 0; i--)
-		{
-			_target[i-1] = ((uint8_t*)source)[i-1];
-		}		
-	}
-	return dest;
-}
