--- kloader/memset.c
@@ -1,12 +1,0 @@
-
-#include "cstub.h"
-
-void *memset(void *s, int value, size_t length)
-{
-	uint8_t *_target = (uint8_t*)s;
-	for (size_t i = 0; i < length; i++)
-	{
-		_target[i] = (uint8_t) value;
-	}
-	return s;
-}
