+++ kloader/pmm.cpp
@@ -1,0 +1,10 @@
+
+#include "pmm.hpp"
+
+uint8_t PMM::Bitmap[BITMAP_SIZE];
+	
+
+retval_t PMM::Init(multiboot2_info_t *mbi) {
+
+	return RETVAL_OK;
+}
